+++ packages/celer-web-app/src/core/map/index.ts
@@ -1,7 +1,5 @@
 export * from "./MapEngine";
-export * from "./MapCore";
 export * from "./convert";
 export * from "./type";
-export * from "./pconvert";
 export * from "./tileset";
 export * from "./MapCanvas";